+++ cypress/e2e/privatePaths/render/deployment/dolt/backups.spec.ts
@@ -1,10 +1,10 @@
 import { desktopDevicesForAppLayout } from "@utils/devices";
 import { runTestsForDevices } from "@utils/index";
-import { doltgresConfig } from "../shared/deploymentConfigs";
+import { doltConfig } from "../shared/deploymentConfigs";
 import { backupsTests } from "../shared/deploymentTestBuilders";
 
-const pageName = "Doltgres deployment backups page";
-const { ownerName, depName } = doltgresConfig;
+const pageName = "Deployment backups page";
+const { ownerName, depName } = doltConfig;
 const currentPage = `/deployments/${ownerName}/${depName}?tab=backups`;
 const loggedIn = true;
 const skip = false;
@@ -13,7 +13,7 @@ describe(pageName, () => {
   // TODO: Fix for mobile
   const devices = desktopDevicesForAppLayout(
     pageName,
-    backupsTests(doltgresConfig),
+    backupsTests(doltConfig),
   );
   runTestsForDevices({ currentPage, devices, skip, loggedIn });
 });
