+++ cypress/e2e/privatePaths/render/deployment/dolt/connectivity.spec.ts
@@ -1,15 +1,15 @@
 import { allDevicesForAppLayout } from "@utils/devices";
 import { runTestsForDevices } from "@utils/index";
-import { doltgresConfig } from "../shared/deploymentConfigs";
+import { doltConfig } from "../shared/deploymentConfigs";
 import { connectivityTests } from "../shared/deploymentTestBuilders";
 
-const pageName = "Doltgres deployment database page, connectivity section";
-const { ownerName, depName } = doltgresConfig;
+const pageName = "Deployment database page, connectivity section";
+const { ownerName, depName } = doltConfig;
 const currentPage = `/deployments/${ownerName}/${depName}?tab=connectivity`;
 const loggedIn = true;
 
 describe(pageName, () => {
-  const tests = connectivityTests(doltgresConfig);
+  const tests = connectivityTests(doltConfig);
   const devices = allDevicesForAppLayout(pageName, tests, tests);
   runTestsForDevices({ currentPage, devices, loggedIn });
 });
