+++ cypress/e2e/privatePaths/render/deployment/dolt/index.spec.ts
@@ -1,18 +1,18 @@
 import { allDevicesForAppLayout } from "@utils/devices";
 import { runTestsForDevices } from "@utils/index";
-import { doltgresConfig } from "../shared/deploymentConfigs";
+import { doltConfig } from "../shared/deploymentConfigs";
 import { indexTests } from "../shared/deploymentTestBuilders";
 
-const pageName = "Doltgres deployment page";
-const { ownerName, depName } = doltgresConfig;
+const pageName = "Deployment page";
+const { ownerName, depName } = doltConfig;
 const currentPage = `/deployments/${ownerName}/${depName}`;
 const loggedIn = true;
 
 describe(pageName, () => {
   const devices = allDevicesForAppLayout(
     pageName,
-    indexTests(doltgresConfig),
-    indexTests(doltgresConfig, true),
+    indexTests(doltConfig),
+    indexTests(doltConfig, true),
     false,
     true,
   );
