+++ cypress/e2e/privatePaths/render/deployment/dolt/monitoring.spec.ts
@@ -1,18 +1,18 @@
 import { allDevicesForAppLayout } from "@utils/devices";
 import { runTestsForDevices } from "@utils/index";
-import { doltgresConfig } from "../shared/deploymentConfigs";
+import { doltConfig } from "../shared/deploymentConfigs";
 import { monitoringTests } from "../shared/deploymentTestBuilders";
 
-const pageName = "Doltgres deployment monitoring page";
-const { ownerName, depName } = doltgresConfig;
+const pageName = "Deployment monitoring page";
+const { ownerName, depName } = doltConfig;
 const currentPage = `/deployments/${ownerName}/${depName}?tab=monitoring`;
 const loggedIn = true;
 
 describe(pageName, () => {
   const devices = allDevicesForAppLayout(
     pageName,
-    monitoringTests(doltgresConfig),
-    monitoringTests(doltgresConfig, true),
+    monitoringTests(doltConfig),
+    monitoringTests(doltConfig, true),
   );
   runTestsForDevices({ currentPage, devices, loggedIn });
 });
