+++ cypress/e2e/privatePaths/render/deployment/dolt/settings.spec.ts
@@ -3,18 +3,18 @@ import {
   iPhoneXForAppLayout,
 } from "@utils/devices";
 import { runTestsForDevices } from "@utils/index";
-import { doltgresConfig } from "../shared/deploymentConfigs";
+import { doltConfig } from "../shared/deploymentConfigs";
 import { settingsTests } from "../shared/deploymentTestBuilders";
 
-const pageName = "Doltgres deployment settings page";
-const { ownerName, depName } = doltgresConfig;
+const pageName = "Deployment settings page";
+const { ownerName, depName } = doltConfig;
 const currentPage = `/deployments/${ownerName}/${depName}?tab=settings`;
 const loggedIn = true;
 
 describe(pageName, () => {
   const devices = [
-    ...desktopDevicesForAppLayout(pageName, settingsTests(doltgresConfig)),
-    iPhoneXForAppLayout(pageName, settingsTests(doltgresConfig, true)),
+    ...desktopDevicesForAppLayout(pageName, settingsTests(doltConfig)),
+    iPhoneXForAppLayout(pageName, settingsTests(doltConfig, true)),
   ];
   runTestsForDevices({ currentPage, devices, loggedIn });
 });
